[Fix] change password
