+++ src/main/java/com/plus1250/jobaTrend/model/dto/UserDTO.java
@@ -24,6 +24,10 @@ public class UserDTO {
         this.nickName = nickName;
     }
 
+    public UserDTO(String email) {
+        this.email = email;
+    }
+
     public UserDTO(String password, String email) {
         this.password = password;
         this.email = email;
