+++ src/main/java/com/plus1250/jobaTrend/repository/UserRepository.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, String> {
     Optional<User> findByEmail(String email);
 
+
     User save(User saveUser);
 
     User findByUserId(String email);
